Improve handling of quotes
Detect when a quote is selected and a new quote key is pressed, replacing the selected quote with new one instead of surrounding that quote with new quotes. Additionally, check whether that quote is used to form string literal (assuming that it is a single-line string literal) and replace paired quote for that literal too.